Moved code around, mostly under pkg/automate
