+++ cmd/tools/automate/get-microservices-metadata.go
@@ -9,24 +9,16 @@ import (
 	"github.com/spf13/viper"
 	"k8s.io/client-go/kubernetes"
 
+	"github.com/dolittle/platform-api/pkg/platform"
+	"github.com/dolittle/platform-api/pkg/platform/automate"
 	"k8s.io/client-go/tools/clientcmd"
 )
-
-type MicroserviceMetadataShortInfo struct {
-	CustomerID       string `json:"customerId"`
-	CustomerName     string `json:"customerName"`
-	ApplicationID    string `json:"applicationId"`
-	ApplicationName  string `json:"applicationName"`
-	Environment      string `json:"environment"`
-	MicroserviceID   string `json:"microserviceId"`
-	MicroserviceName string `json:"microserviceName"`
-}
 
 var getMicroservicesMetaDataCMD = &cobra.Command{
 	Use:   "get-microservices-metadata",
 	Short: "Get all microservices metadata from the cluster",
 	Long: `
-	go run main.go tools get-microservices-metadata
+	go run main.go tools automate get-microservices-metadata
 
 	Outputs:
 		TODO
@@ -49,14 +41,14 @@ var getMicroservicesMetaDataCMD = &cobra.Command{
 			panic(err.Error())
 		}
 
-		microservices, err := getAllCustomerMicroservices(ctx, client)
+		microservices, err := automate.GetAllCustomerMicroservices(ctx, client)
 		if err != nil {
 			panic(err.Error())
 		}
 
-		data := make([]MicroserviceMetadataShortInfo, 0)
+		data := make([]platform.MicroserviceMetadataShortInfo, 0)
 		for _, microservice := range microservices {
-			data = append(data, MicroserviceMetadataShortInfo{
+			data = append(data, platform.MicroserviceMetadataShortInfo{
 				ApplicationID:    microservice.Application.ID,
 				ApplicationName:  microservice.Application.Name,
 				Environment:      microservice.Environment,
