+++ pkg/platform/microservice/k8s/doc.go
@@ -7,6 +7,7 @@ import (
 	"log"
 
 	"github.com/dolittle/platform-api/pkg/dolittle/k8s"
+	"github.com/dolittle/platform-api/pkg/platform/automate"
 	v1 "k8s.io/api/apps/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -32,31 +33,10 @@ func CreateTodoIngress() k8s.Ingress {
 	}
 }
 
-// K8sGetDeployments gets all microservice deployments in the given namespace.
-// A microservice deployment is a deployment with the 'dolittle.io/microservice-id' annotation and the 'microservice' label set
-func K8sGetDeployments(client kubernetes.Interface, context context.Context, namespace string) ([]v1.Deployment, error) {
-	deployments, err := client.AppsV1().Deployments(namespace).List(context, metaV1.ListOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	var microserviceDeployments []v1.Deployment
-	for _, deployment := range deployments.Items {
-		if _, ok := deployment.Labels["microservice"]; !ok {
-			continue
-		}
-		if _, ok := deployment.Annotations["dolittle.io/microservice-id"]; !ok {
-			continue
-		}
-		microserviceDeployments = append(microserviceDeployments, deployment)
-	}
-	return microserviceDeployments, nil
-}
-
 // Gets the deployment that is linked to the microserviceID in the given namespace
 // Caveat: This will only get the first deployment that has the given dolittle.io/microservice-id annotation
 func K8sGetDeployment(client kubernetes.Interface, context context.Context, namespace, microserviceID string) (v1.Deployment, error) {
-	deployments, err := K8sGetDeployments(client, context, namespace)
+	deployments, err := automate.GetDeployments(context, client, namespace)
 	if err != nil {
 		return v1.Deployment{}, err
 	}
